pyproject.toml
[project]
name = "code-explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
